Implement LP Token Lock
This code was adapted from CORE contracts. Some minor edits were needed, as the CORE contracts have their token pair defined in their LGE where as CYCLE will not have an LGE. In the CORE contract, the code used is:

In the transfer function:

(At the beginning of the function) 
        uint _LPSupplyOfCYCLEETHPairTotal = IERC20(tokenPair).totalSupply();
        if(sender == tokenPair) {
            require(lastTotalSupplyOfCYCLEETHLPTokens <= _LPSupplyOfCYCLEETHPairTotal, "Liquidity withdrawals forbidden for CYCLE/ETH pair");
        }

TRANSFER CODE HERE

(At the very end of the transfer function)
        lastTotalSupplyOfDFGETHLPTokens = _LPSupplyOfDFGETHPairTotal;


Calling the sync function is what locks the LP tokens. After calling this, the LP will no longer be unwrappable. 

    function sync() internal onlyOwner {
        uint _LPSupplyOfDFGETHPairTotalPairTotal = IERC20(tokenUniswapPairDFGETH).totalSupply();
        lastTotalSupplyOfDFGETHLPTokens = _LPSupplyOfDFGETHPairTotalPairTotal;
    }